requirements.txt
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
build==1.4.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.102.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
fastapi==0.135.1
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
googleapis-common-protos==1.73.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.7.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kubernetes==35.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
mmh3==5.2.1
mpmath==1.3.0
narwhals==2.18.0
numpy==2.4.3
oauthlib==3.3.1
onnxruntime==1.24.4
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
prettytable==3.17.0
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
skops==0.13.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlglot==30.4.3
sqlparse==0.5.5
starlette==1.0.0
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
zipp==3.23.0
